Debugging thread_unit_lib.  Phase 2: I think it works.

+++ my_pthread_t.h
@@ -36,7 +36,7 @@
 */
 typedef struct my_pthread_t {
 
-    unsigned int        threadID;
+    long int 	threadID;
 	void *              return_val;
 	int                 priority;
